GesturesDataManager class was implemented

+++ Assets/Scripts/GesturesSystem/Point.cs
@@ -4,11 +4,11 @@ namespace GesturesSystem
 {
 	public class Point
 	{
-		public ulong ID { get; private set; }
+		public int ID { get; private set; }
 		public float2 Position { get; private set; }
 		public int2 LookUpTablePosition { get; private set; }
 
-		public Point (ulong id, float2 position)
+		public Point (int id, float2 position)
 		{
 			ID = id;
 			Position = position;
